refactor: src/lib 收编完成 + dsh_install 退役 + install→bootstrap 改名
- tools/lib 并入 src/lib（state/config/protocol/wake），src/tools 只留工具壳
  （dsh-session/approve/run/cancel），工具 import ../lib/*；src/lib 为唯一 lib 层
- dsh_install 工具退役（D6 零干预后安装由自动链 + Bootstrap 自举承担）：删
  tools/dsh-install.js + skills/dsh-install + index.js 注册（HANAKO_TOOLS 收敛为
  [dshSession, dshApprove]）；能力层 installDepsFromPlugin/verifyDepsSmoke 保留
- lib/install.js → lib/bootstrap.js（名实相符：主体服务 Bootstrap 自动链）
build + 27/27 通过。
