+++ src/index.js
@@ -35,15 +35,13 @@ import {
   ERROR_CLASS_GUIDANCE,
 } from "./lib/errclass.js";
 // 5 个工具模块（ESM 导出 name/description/parameters/execute(+sessionPermission)；
-// dsh_update 已并入 dsh_install 四合一，见 tools/dsh-install.js）
-import * as dshInstall from "./tools/dsh-install.js";
 import * as dshApprove from "./tools/dsh-approve.js";
 // T7e 工具收敛：dsh-run / dsh-cancel 并入 dsh-session（create/send/cancel 分支），
 // 不再作为独立工具注册；dsh-session.js 内部 import 复用它们的 execute 实现。
 import * as dshSession from "./tools/dsh-session.js";
 // 宿主 task 体系接入（handler.abort → dsh session.cancel）：取消链路收归宿主 task
 // 协议（task:abort → handler.abort → session.cancel），dsh_cancel 不再直连取消
-import { callUnaryBus } from "./tools/lib/protocol.js";
+import { callUnaryBus } from "./lib/protocol.js";
 // 路由工厂（默认导出）
 import registerWebuiRoutes from "./routes/webui.js";
 import registerCardRoutes from "./routes/card.js";
@@ -51,7 +49,7 @@ import registerCardRoutes from "./routes/card.js";
 // 工具清单（registerTool 消费普通契约；宿主自动加 pluginId_ 前缀）
 // T7e：dsh-run/dsh-cancel 并入 dsh-session（create/send/cancel）；approve 独立保留
 // （权限应答语义正交，非会话操作）；install 独立（依赖管理）。
-const HANAKO_TOOLS = [dshSession, dshInstall, dshApprove];
+const HANAKO_TOOLS = [dshSession, dshApprove]; // dsh_install 已退役：D6 零干预后安装由自动链 + Bootstrap 自举承担（2026-09-04）
 
 // ---- 统一日志（时间戳会话文件；旧日志 zstd 压缩保留经 migrate.js 统一迁移入口）----
 // DSHana 插件全量运行日志：每次插件会话创建 <YYYYMMDD-HHmmss-SSS>.log 真实文件（文件名
